Detect cycles in inheritance graph during completion
If a cycle is detected,
don't return any of the properties or methods of the class.
Prevent ResolvedJavaTypeInfo from being fetched from qute.jdt multiple times.
This prevents completion from freezing when you try to complete the
properties and methods of an object whose class is part of an
inheritance cycle.

Closes #725

Signed-off-by: David Thompson <davthomp@redhat.com>